p4i integration Integrate 177514 from 4.0 to main: Fixes:    Store the size as 64bit Task:     71983 RevBy:    Andy
[git-p4: depot-paths = "//depot/qt/main/": change = 177515]

+++ src/corelib/io/qfileinfo.cpp
@@ -56,7 +56,7 @@ public:
         mutable QHash<int, QString> fileNames;
 
         mutable uint cache_enabled : 1;
-        mutable uint fileSize;
+        mutable qint64 fileSize;
         mutable QDateTime fileTimes[3];
         mutable uint fileFlags;
         mutable uchar cachedFlags;
